fix draft js editor image rendering
